Give better error message if Nano SVG is unavailable, if possible
Getting errors due to "missing" Nano SVG headers can be confusing, as
it's not immediately clear why are they missing, so try to make things
more explicit by using __has_include(), if available, to test for the
header before including it and giving a better error message if it
fails.

It still won't help people not using C++17 or those not reading error
messages, but should be better than nothing.