Revert "Merge branch 'ak_high_dpi' of https://github.com/kosh543/wxWidgets"
This reverts commit b784da19554b50178d73c6aef45328c21d1f5193, reversing
changes made to df898f3504ecf284ec3d7151fd7f382ef2c4856b.

Undo the addition of wxBitmapBundle support to XRC, as it is finally
going to be done in a different way soon -- but for now just remove it
to ensure that nobody starts using it before it's finalized.

See #2633.

Closes #22014.