Add wxWindow::FromPhys() and ToPhys() functions
Provide these functions for consistency with the existing FromDIP() and
ToDIP().

We also can use ceil() (rather than round() used by wxSize operators) in
FromPhys() to ensure that we never truncate contents of a physical
bitmap, which allows to replace the existing wxStaticBitmap code with
just a call to this function.