Fix applying scale in wrong direction in wxOSXImageFromBitmap()
We need to use the bitmap size in logical pixels for NSImage and we
definitely must not multiply the physical size by the scale factor, only
dividing it makes sense -- so just do it, by using GetScaledXXX()
instead of GetXXX() when creating the NSImage to use for wxBitmap.

Surprisingly, this seems to have been broken ever since the addition of
this code in fbce6997ab (first part of OSX adaptations, 2021-09-28)
without anybody noticing it until now.

Closes #22019.