Fix determining DPI info for wxDialog creation
WM_NCCREATE is not received for dialogs, so use a different message.
There seem to be no other messages that are always and only send on
creation, so use WM_NCCALCSIZE which seems always generated but not too
often.

Use m_activeDPI to determine if the DPI variables have been initialized
or not, instead of adding another variable for this.

Closes #22133.