Use ccache in MSW cross build CI workflows too
This should hopefully help to speed up this build as well.

Unfortunately we can't easily use ccache for the Mac builds because this
requires having brew installed and we don't have it currently.

We probably can use it for CMake workflow but this probably requires
some arcane way of configuring CMake to use ccache that I'm not aware
about and would need to be done by someone else.