Don't show value-less wxDataViewVirtualListModel cells in wxGTK
For some reason, calls to wxGtkTreeSetVisibleProp() were skipped when
using virtual list model in wxGTK implementation, resulting in showing
the value of the previous (i.e. upper) cell for the rows of this model
for which no value was available.

Simply remove IsVirtualListModel() checks and always set the cell
visibility to fix this.

This commit is best viewed ignoring whitespace-only changes.