Implement wxBitmapBundle::GetBitmap() on macOS
Implement wxOSXImageBundleImpl::GetBitmap() to return the underlying
NSImage. This ignores size, because NSImage - and so wxBitmap - already
is multi-resolution representation.

This un-breaks loading bundle bitmaps via wxArtProvider, which
previously used NSImage to load multi-resolution bitmap into wxBitmap,
but lost the ability to return simple bitmaps in the wxBitmapBundle
conversion.

See #22449.