Fix determining the size of wxBitmapComboBox bitmaps
Instead of using the size of the first item bitmap, use the size best
suited for all the bitmaps, which may result in better appearance if the
different bitmaps are not all available in the same sizes.

This also fixes the unit test after c8e119f67b (Fix margin between
wxBitmapComboBox images and text in high DPI, 2022-05-25) and should
have been part of it.