Don't update wxBitmapComboBox bitmaps size if there are none
Correct the changes of c7bafb3982 (Fix determining the size of
wxBitmapComboBox bitmaps, 2022-05-25) and avoid changing m_usedImgSize
if we have no bitmaps at all, as it happens when adding the very first
bitmap: in this case, we set m_usedImgSize in our OnAddBitmap() before
actually adding the bitmap to m_bitmapbundles in DoInsertItems() called
from the base class Append() or Insert(), and wxMSW implementation calls
RecreateControl() from its overridden OnAddBitmap() which results in a
call to UpdateInternals() in between them.

All this is rather confusing and it would arguably be better to avoid
setting m_usedImgSize until the bitmap is actually added to
m_bitmapbundles, but this would require further changes, so for now just
avoid setting m_usedImgSize to (0, 0).