Fix confusion between width and height in wxArtProvider
Amazingly, the fact that the bitmap height was compared with the
required width went unnoticed even though it was there since this code
was added back in e53a95bcb1 (Applied patch for ArtProvider.,
2005-03-12) and even survived migration to a different file in
48c08906de (Move bitmap resizing from wxDefaultArtProvider to base
class, 2022-02-05).

That this has never been a problem seems to conclusively prove that
non-square icons simply don't exist, but still fix the comparisons just
to be tidy.