Allow upscaling wxArtProvider bitmaps by integer scale factor
While upscaling the bitmaps by factors smaller than 2 results in truly
awful results, upscaling them by 2, or 3, is better on high DPI screens
than using tiny bitmaps that are difficult to read in high resolution,
so do scale them up in this case.

This finally makes wxBitmapBundleImplArt::GetBitmap() work as expected,
as it now returns the upscaled bitmap rather than the original small
bitmap in the middle of a bigger canvas.