Fix refresh after calling generic wxListCtrl::EnsureVisible()
The control still needs to be repainted after calling
RecalculatePositions(), so don't reset the dirty flag in it and do it
only in RecalculatePositionsAndRefresh() instead.

Arguably, EnsureVisible() should call the latter function itself, but
don't change this for now and just correct the regression caused by the
changes of 75165377b1 (Avoid bool argument in wxListMainWindow::
RecalculatePositions(), 2021-08-06).

Closes #22511.