Fix calling TransferDataFromCustomControls() with common dialogs
When using common dialogs, because IFileDialog-based implementation is
not available either at compile- or run-time, this function needs to be
called while the extra controls still exist, i.e. before ShowModal()
returns, so do it from CDN_FILEOK handler.

Move the code that called it previously into the new MSWOnFileOK() for
consistency with the other callbacks.

Closes #22521.