Fix creating IFileDialog-based "Save" file dialogs
We need IFileDialog and not an IFileOpenDialog when creating a generic
file dialog, all the code had been already updated to only use the
former and not the latter, except for CoCreateInstance() itself which
still always asked for the latter, resulting in a failure when trying to
obtain IFileOpenDialog from a "Save" file dialog not providing it.

Also update some log messages to use the generally correct interface.

See #22521.