Destroy global wxConfig even if main loop didn't run
Move the code freeing the global wxConfig object from wxApp::OnExit(),
which is called only after running the main loop, to wxApp::CleanUp(),
which is always called.

This fixes memory leaks when not running the main loop at all, as e.g.
when showing a modal dialog instead of doing it.