Determine codeset properly
In case of the system default locale the codeset can't be determined in the constructor, because the requested locale is actually not yet set up, so that typically the C locale is queried for the codeset resulting in a usually wrong value. Therefore the codeset has to be determined in `wxUILocaleImplUnix::Use()` (again).