Disable wxWebRequest tests in FreeBSD CI due to pip problems
The situation with Python under FreeBSD is pretty weird, with system
Python being 3.7, but pip only available as py38-pip package and,
unsurprisingly, not working with python3.

This could probably be resolved somehow, but for now I just don't have
enough energy to fight with Python again, so simply don't use it.