Fix windows becoming smaller than best size on DPI change
Update the TLW decoration size as soon as DPI changes to ensure that
GetBestSize() returns correct value: previously it was too small when
moving the window from normal to high DPI, for example, because
GetBestSize() added the smaller, old decoration size to the best
computed client size.

And now that it works correctly, ensure that we never set the window
size to a size smaller than its best size to prevent window elements
from not fitting into the window after the DPI change.

See #18649.