Disable custom Doxygen theme support for CHM generation
Unsurprisingly, this doesn't work at all with CHM, so allow disabling
using the custom theme when doing it by adding more custom variables.

Using (multiple) variables for this is rather ugly, but we already use
them for other options, so it seems better to keep adding them rather
than using some other mechanism.

Adding a different custom header file is even uglier, but it seems to be
the only way to avoid including the theme JS files into CHM.