Add wxDEPRECATED_EXPORT_CORE() macro for wxTransformMatrix
wxDEPRECATED_MSG() and WXDLLIMPEXP_CORE can't be used together in the
same declaration when the former uses the standard attribute ([[...]])
and the latter uses a legacy one (__attribute__((....))), at least not
with gcc 12.

Work around this problem by defining a special new macro that combines
both attributes in a working way.

This is rather ugly, as it would seem to be better to just always define
WXDLLIMPEXP_CORE using the standard attribute, but unfortunately this
doesn't work as the standard attribute must be placed before the
function and variable declarations, while we currently use our DLL
export macros in the middle of the declaration. Maybe we can change all
the declarations doing this later, but for now this is the simplest
solution to the immediate problem.

Note that this commit, unlike its master equivalent, includes the check
for WXBUILDING when defining wxDEPRECATED_EXPORT_CORE(), i.e. it's a
combination of 2d939a36653a7ea015c82c841a30dfd22f90cdad and
c8af0a3b7163bd34e2e8a048baa2f7732593ddd1 from master and not just the
former one.

See #22790, #22820.